pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "nash-mcp"
version = "0.1.13"
description = "Nash MCP"
authors = ["Nash Team <team@nash.run>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nash_mcp", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
requests = ">=2.32.3,<3.0.0"
pandas = ">=2.2.3,<3.0.0"
matplotlib = ">=3.10.0,<4.0.0"
beautifulsoup4 = ">=4.13.3,<5.0.0"
sqlalchemy = ">=2.0.38,<3.0.0"
scikit-learn = ">=1.6.1,<2.0.0"
torch = ">=2.6.0,<3.0.0"
tensorflow = ">=2.18.0,<3.0.0"
html2text = ">=2024.2.26,<2025.0.0"
selenium = ">=4.29.0,<5.0.0"
playwright = ">=1.50.0,<2.0.0"
webdriver-manager = ">=4.0.2,<5.0.0"
openpyxl = ">=3.1.5,<4.0.0"
slugify = ">=0.0.1,<0.0.2"
psycopg2-binary = "^2.9.10"
tabulate = "^0.9.0"
seaborn = "^0.13.2"
statsmodels = "^0.14.4"
python-dotenv = "^1.0.1"
python-docx = "^1.1.2"
pypdf2 = "^3.0.1"
pypdf = "^5.3.1"
docx2txt = "^0.8"
ebooklib = "^0.18"
xlrd = "^2.0.1"
xlsxwriter = "^3.2.2"
pyxlsb = "^1.0.10"
opencv-python = "^4.11.0.86"
pydub = "^0.25.1"
librosa = "^0.11.0"
mutagen = "^1.47.0"
av = "^14.2.0"
ffmpeg-python = "^0.2.0"
rarfile = "^4.2"
py7zr = "^0.22.0"
zipfile38 = "^0.0.3"
trimesh = "^4.6.4"
pygltflib = "^1.16.3"
pywavefront = "^1.3.3"
pydicom = "^3.0.1"
python-pptx = "^1.0.2"
psd-tools = "^1.10.7"
pdf2image = "^1.17.0"
gitpython = "^3.1.44"
pymongo = "^4.11.2"
construct = "^2.10.70"
pyelftools = "^0.32"
h5py = "^3.13.0"
netcdf4 = "^1.7.2"
biopython = "^1.85"
twisted = "^24.11.0"
geckodriver-autoinstaller = "^0.1.0"
chromedriver-autoinstaller = "^0.6.4"
async-timeout = "^5.0.1"
tenacity = "^9.0.0"
retry = "^0.9.2"
browser-use = "^0.1.40"
langchain-openai = "0.3.1"
langchain-anthropic = "0.3.3"
psutil = "^5.9.8"
mcp = {extras = ["cli"], version = "^1.4.1"}

[tool.poetry.scripts]
nash = "nash_mcp.server:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^6.0.0"
black = "^25.1.0"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 120
